Content Manager codes

+++ src/main/java/ng/hotsystems/contentManager/services/ArticleServiceImpl.java
@@ -13,6 +13,8 @@ import ng.hotsystems.contentManager.utils.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
+
 @Service
 public class ArticleServiceImpl implements ArticleService {
 
@@ -24,7 +26,9 @@ public class ArticleServiceImpl implements ArticleService {
     @Override
     public Article addArticle(AddArticleRequest request) {
         Article foundArticle = articleRepository.findArticleByTitle(request.getTitle());
-        if (foundArticle != null) throw new ArticleExistsException("This article already exists. Write a new one?");
+        if (foundArticle != null && Objects.equals(foundArticle.getBlogName(), request.getBlogName())) {
+            throw new ArticleExistsException("This article already exists. Write a new one?");
+        }
 
         Article newArticle = new Article();
         Mapper.map(request, newArticle);
@@ -36,26 +40,32 @@ public class ArticleServiceImpl implements ArticleService {
     public void deleteArticle(DeleteArticleRequest deleteRequest) {
         Article article = articleRepository.findArticleByTitle(deleteRequest.getTitle());
         if (article == null) throw new ArticleDoesNotExistException("This article does not exist.");
-        for (Comment comment: article.getComments()) {
-            commentService.deleteComment(comment);
-        }
+        else if (Objects.equals(article.getBlogName(), deleteRequest.getBlogName())) {
+            for (Comment comment : article.getComments()) {
+                commentService.deleteComment(comment);
+            }
 
-        articleRepository.delete(article);
+            articleRepository.delete(article);
+        }
     }
 
     @Override
     public Article viewArticle(FindArticleRequest request) {
         Article article = articleRepository.findArticleByTitle(request.getTitle());
         if (article == null) throw new ArticleDoesNotExistException("This article does not exist.");
-
-        return articleRepository.findArticleByTitle(request.getTitle());
+        else if (Objects.equals(article.getBlogName(), request.getBlogName())) {
+            return article;
+        }
+        return null;
     }
 
     @Override
     public Article addComment(AddCommentRequest request) {
         Comment newComment = commentService.addComment(request);
         Article foundArticle = articleRepository.findArticleByTitle(request.getArticleTitle());
-        foundArticle.getComments().add(newComment);
+        if (Objects.equals(foundArticle.getBlogName(), request.getBlogName())) {
+            foundArticle.getComments().add(newComment);
+        }
 
         return articleRepository.save(foundArticle);
     }
